Move bullhorn script to report.js
